Refactor class names etc
