+++ EventLook/View/LogPickerWindow.xaml
@@ -1,0 +1,34 @@
+﻿<Window x:Class="EventLook.View.LogPickerWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:EventLook.View"
+        xmlns:viewmodel="clr-namespace:EventLook.ViewModel"
+        d:DataContext="{d:DesignInstance Type=viewmodel:LogPickerViewModel}"
+        mc:Ignorable="d"
+        Title="Open Event Log" Height="450" Width="800">
+    <Grid>
+        <Grid.Resources>
+            <Style TargetType="{x:Type Button}">
+                <Setter Property="Width" Value="60" />
+                <Setter Property="Height" Value="22" />
+                <Setter Property="Margin" Value="5,0,0,0" />
+            </Style>
+        </Grid.Resources>
+        <Grid.RowDefinitions>
+            <RowDefinition />
+            <RowDefinition Height="Auto" />
+        </Grid.RowDefinitions>
+        <StackPanel Orientation="Vertical" Grid.Row="0">
+            <ListBox x:Name="listBox1" SelectionMode="Single" MaxHeight="370"
+                     ItemsSource="{Binding LogChannels, Mode=OneWay}" DisplayMemberPath="Path"
+                     SelectedItem="{Binding SelectedChannel}"
+                     MouseDoubleClick="ListBox_MouseDoubleClick"/>
+        </StackPanel>
+        <StackPanel Grid.Row="1" Orientation="Horizontal" HorizontalAlignment="Right" Margin="10,10">
+            <Button Content="OK" Margin="10,0" Click="OkButton_Click" IsDefault="True"/>
+            <Button Content="Cancel" Margin="10,0" IsCancel="True"/>
+        </StackPanel>
+    </Grid>
+</Window>
