+++ EventLook/View/MainWindow.xaml
@@ -36,14 +36,14 @@
     <Window.InputBindings>
         <KeyBinding Key="F5" Command="{Binding RefreshCommand}"/>
         <KeyBinding Key="F5" Modifiers="Shift" Command="{Binding CancelCommand}"/>
-        <KeyBinding Key="L" Modifiers="Ctrl" Command="{Binding OpenLocalCommand}"/>
+        <KeyBinding Key="L" Modifiers="Ctrl" Command="{Binding OpenLogPickerCommand}"/>
         <KeyBinding Key="O" Modifiers="Ctrl" Command="{Binding OpenFileCommand}"/>
     </Window.InputBindings>
     
     <DockPanel>
         <Menu DockPanel.Dock="Top" Background="White">
             <MenuItem Header="_File">
-                <MenuItem Header="Open Event _Log" InputGestureText="Ctrl+L" Command="{Binding OpenLocalCommand}"/>
+                <MenuItem Header="Open Event _Log" InputGestureText="Ctrl+L" Command="{Binding OpenLogPickerCommand}"/>
                 <MenuItem Header="_Open .evtx File" InputGestureText="Ctrl+O" Command="{Binding OpenFileCommand}"/>
                 <Separator />
                 <MenuItem Header="_Launch Windows Event Viewer" Command="{Binding LaunchEventViewerCommand}"/>
