--- EventLook/View/OpenLocalLogWindow.xaml
@@ -1,28 +1,0 @@
-﻿<Window x:Class="EventLook.View.OpenLocalLogWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:EventLook.View"
-        xmlns:viewmodel="clr-namespace:EventLook.ViewModel"
-        d:DataContext="{d:DesignInstance Type=viewmodel:OpenLocalLogViewModel}"
-        mc:Ignorable="d"
-        Title="OpenLocalLogWindow" Height="450" Width="800">
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="380" />
-            <RowDefinition Height="20" />
-        </Grid.RowDefinitions>
-        <StackPanel Orientation="Vertical" Grid.Row="0">
-            <TextBlock Text="All logs"/>
-            <ListBox x:Name="listBox1" SelectionMode="Single" MaxHeight="380"
-                     ItemsSource="{Binding LogChannels, Mode=OneWay}" DisplayMemberPath="Path"
-                     SelectedItem="{Binding SelectedChannel}"
-                     MouseDoubleClick="ListBox_MouseDoubleClick"/>
-        </StackPanel>
-        <StackPanel Orientation="Horizontal" Grid.Row="1">
-            <Button Content="OK" Click="OkButton_Click" IsDefault="True"/>
-            <Button Content="Cancel" IsCancel="True"/>
-        </StackPanel>
-    </Grid>
-</Window>
