+++ EventLook/ViewModel/LogPickerViewModel.cs
@@ -1,0 +1,24 @@
+﻿using EventLook.Model;
+using System;
+using System.Collections.Generic;
+using System.Diagnostics.Eventing.Reader;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace EventLook.ViewModel;
+
+public class LogPickerViewModel
+{
+    public LogPickerViewModel()
+    {
+        var session = new EventLogSession();
+        LogChannels = new List<LogChannel>();
+        foreach (var channelName in session.GetLogNames())
+        {
+            LogChannels.Add(new LogChannel { Path = channelName });
+        }
+    }
+    public List<LogChannel> LogChannels { get; set; }
+    public LogChannel SelectedChannel { get; set; }
+}
